python_backend/requirements.txt
discord.py
sqlalchemy
solana
solders
python-dotenv
aiohttp
